projects/bridgit-matching-engine/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.6.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0.1
pytest>=7.4.0
# Optional:
# faiss-cpu>=1.7.4
# opensearch-py>=2.4.2
# openai>=1.46.0
# boto3>=1.34.162
